requirements.txt
streamlit
sentence-transformers
weaviate-client>=4.0.0
python-dotenv
tavily-python
